fixed so that query data works
