+++ frontend/package.json
@@ -51,6 +51,7 @@
     "react-fast-marquee": "^1.6.2",
     "recharts": "^2.12.7",
     "sass": "^1.69.7",
+    "sql.js": "^1.10.3",
     "styled-components": "^6.1.8",
     "tailwind-merge": "^2.3.0",
     "tippy.js": "^6.3.7",
